SDCT/src/python/requirements.txt
pyaudio==0.2.14
python-socketio==5.11.2
pygame==2.5.2
setuptools
requests
